delivary page some changes

+++ WindowsFormsApplication1/Purchase Delivery.cs
@@ -1368,8 +1368,8 @@ namespace WindowsFormsApplication1
                         int quent1 = Convert.ToInt32(quent);
                         // int curentQuntity = Convert.ToInt32(que);
                         //int quntity = curentQuntity - quent1;
-                        int cuentQuantity = Convert.ToInt32(currid);
-                        int lastQuantity = cuentQuantity + quent1; //cuentQuantity - curentQuntity;
+                       Double cuentQuantity = Convert.ToDouble(currid);
+                       Double lastQuantity = cuentQuantity + quent1; //cuentQuantity - curentQuntity;
                         //int resivquenty = lastQuantity + quent1;
                         //string currid1 = resivquenty.ToString();
                         string updateQurry = "update ItemQuantityDetail set CurrentQuantity='" + lastQuantity + "'where ItemId='" + itid + "'";
